refactor(conect_api): resolve
